+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to models are documented here.
 
-## Unreleased
+## [0.3.24] - 2026-09-02
 
 ### Added